DAO&BO complete. Hibernate optimization.
