+++ src/main/java/com/coldrain/database/models/SongModel.java
@@ -3,14 +3,25 @@ package com.coldrain.database.models;
 import com.coldrain.database.generic.DatabasePOJO;
 import com.coldrain.database.utils.InitSessionFactory;
 import com.coldrain.playlist.generic.MetaSong;
-import org.hibernate.Session;
-import org.hibernate.Transaction;
-import org.hibernate.query.Query;
-
-import javax.persistence.*;
 import java.io.Serializable;
 import java.util.HashSet;
 import java.util.Set;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Index;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
+import javax.persistence.SequenceGenerator;
+import javax.persistence.Table;
+import org.hibernate.Session;
+import org.hibernate.Transaction;
+import org.hibernate.annotations.Fetch;
+import org.hibernate.annotations.FetchMode;
+import org.hibernate.query.Query;
 
 @Entity
 @Table(name = "Song", indexes = {@Index(name = "index_trackTitle",columnList = "trackTitle ASC")})
@@ -27,7 +38,8 @@ public class SongModel extends DatabasePOJO implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "song_seq")
     private Integer songid;
 
-    @OneToMany(mappedBy = "songM")
+    @OneToMany(mappedBy = "songM", fetch = FetchType.LAZY)
+    @Fetch(FetchMode.SELECT)
     private Set<MetaModel> metaModels = new HashSet<MetaModel>();
 
     @ManyToOne
@@ -48,6 +60,8 @@ public class SongModel extends DatabasePOJO implements Serializable {
     public SongModel(MetaSong meta) {
         this.trackTitle = meta.getTrackTitle();
     }
+
+    public SongModel(String trackTitle){ this.trackTitle = trackTitle;}
 
     public Integer getSongid() {
         return songid;
