+++ src/test/java/com/coldrain/ncm/utils/NcmConfigTest.java
@@ -1,10 +1,14 @@
 package com.coldrain.ncm.utils;
 
 import com.coldrain.FirstSpringApp;
+import com.coldrain.database.service.AudioDBService;
+import com.coldrain.exception.MetaIOException;
 import com.coldrain.ncm.NcmService;
 import com.coldrain.ncm.jsonSupp.PlayListBean;
+import com.coldrain.playlist.generic.MetaSong;
 import java.util.List;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
@@ -17,6 +21,7 @@ public class NcmConfigTest {
     static org.springframework.context.ApplicationContext applicationContext;
     static NcmConfig ncmConfig;
     static NcmService ncmService;
+    static AudioDBService audioDBService;
 
     @BeforeAll
     static void initContext(){
@@ -30,12 +35,25 @@ public class NcmConfigTest {
     }
 
     @Test
+    @Disabled
     void test2(){
         ncmService = applicationContext.getBean((NcmService.class));
         List<PlayListBean> plbList = ncmService.getPlaylists_publicTest();
         for(PlayListBean aBean: plbList){
             System.out.println(aBean);
         }
+    }
+
+    @Test
+    void test3() throws MetaIOException {
+        audioDBService = (AudioDBService) applicationContext.getBean("AudioDBService");
+        System.out.println("AudioDBService RETRIEVED");
+        audioDBService.scanSingle(new MetaSong("F:\\CloudMusic\\电台节目\\初弦__ - 「遗尘漫步」PV曲.mp3"));
+    }
+
+    @Test
+    void test4() throws MetaIOException {
+        audioDBService.scanSingle(new MetaSong("F:\\CloudMusic\\电台节目\\初弦__ - 「画中人」 夕.mp3"));
     }
 
 }
